pyproject.toml
[project]
name = "lineage-backend"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.32.1",
    "neo4j>=5.28.1",
    "databricks-sql-connector>=3.4.0",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "pydantic>=2.10.4",
    "apscheduler>=3.10.4",
    "pytz>=2024.1",
    "networkx==3.6.1",
    "rapidfuzz>=3.14.3",
]
